backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pillow
numpy
onnxruntime
